Separated the code into 2 files
